pyproject.toml
[tool.poetry]
name = "poetry-cicd"
version = "0.1.0"
description = ""
authors = ["Gregory Sech <sech.gregory@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
jinja2 = "^3.1.4"
sqlalchemy = "^2.0.31"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
